% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parBM.R
\name{parBM}
\alias{parBM}
\title{Parallel batch means}
\usage{
parBM(chains, r = 1)
}
\description{
Parallel batch means
chain = list of chains, each of equal length
\itemize{
\item r = 1 is regular BM
\item r = 2 is flat-top BM
\item r = 3 is lugsail BM
}
}
