% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parSV.R
\name{parSVE}
\alias{parSVE}
\title{Parallel spectral variance estimator}
\usage{
parSVE(chains, r = 1)
}
\description{
Parallel spectral variance estimator
chain = list of chains, each of equal length
\itemize{
\item r = 1 is regular BM
\item r = 2 is flat-top BM
\item r = 3 is lugsail BM
}
}
