% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_epave.R
\name{sample_epave}
\alias{sample_epave}
\title{Sample EPAVE}
\usage{
sample_epave(
  single_kernel,
  coupled_kernel,
  rinit,
  nmcmc,
  burnin,
  D,
  h,
  state_x_0,
  max_iterations = Inf
)
}
\description{
Sample estimator of asymptotic variance in the CLT for MCMC averages.
}
